move OS-related functions to shell module
