+++ src/utils.ts
@@ -1,15 +1,5 @@
 import assert from "assert"
 import { differenceInMilliseconds } from "date-fns"
-import fs from "fs"
-import path from "path"
-import { promisify } from "util"
-
-import { CommandExecutor, CommandOutput } from "./types"
-
-const fsExists = promisify(fs.exists)
-const fsRmdir = promisify(fs.rmdir)
-const fsMkdir = promisify(fs.mkdir)
-const fsUnlink = promisify(fs.unlink)
 
 export const envVar = (name: string) => {
   const val = process.env[name]
@@ -54,29 +44,6 @@ export const millisecondsDelay = (milliseconds: number) => {
   return new Promise<void>((resolve) => {
     setTimeout(resolve, milliseconds)
   })
-}
-
-export const ensureDir = async (dir: string) => {
-  if (!(await fsExists(dir))) {
-    await fsMkdir(dir, { recursive: true })
-  }
-  return dir
-}
-
-export const removeDir = async (dir: string) => {
-  if (!(await fsExists(dir))) {
-    await fsRmdir(dir, { recursive: true })
-  }
-  return dir
-}
-
-export const initDatabaseDir = async (dir: string) => {
-  dir = await ensureDir(dir)
-  const lockPath = path.join(dir, "LOCK")
-  if (await fsExists(lockPath)) {
-    await fsUnlink(lockPath)
-  }
-  return dir
 }
 
 export const intoError = (value: unknown) => {
@@ -133,98 +100,6 @@ export const escapeHtml = (str: string) => {
     .replace(/'/g, "&#039;")
 }
 
-const walkDirs: (dir: string) => AsyncGenerator<string> = async function* (
-  dir,
-) {
-  for await (const d of await fs.promises.opendir(dir)) {
-    if (!d.isDirectory()) {
-      continue
-    }
-
-    const fullPath = path.join(dir, d.name)
-    yield fullPath
-
-    yield* walkDirs(fullPath)
-  }
-}
-
-export const cleanupProjects = async (
-  executor: CommandExecutor,
-  projectsRoot: string,
-  {
-    includeDirs,
-    excludeDirs = [],
-  }: { includeDirs?: string[]; excludeDirs?: string[] } = {},
-) => {
-  const results: CommandOutput[] = []
-
-  toNextProject: for await (const p of walkDirs(projectsRoot)) {
-    if (!(await fsExists(path.join(p, ".git")))) {
-      continue
-    }
-
-    if (
-      includeDirs !== undefined &&
-      includeDirs.filter((includeDir) => {
-        return isDirectoryOrSubdirectory(includeDir, p)
-      }).length === 0
-    ) {
-      continue toNextProject
-    }
-
-    for (const excludeDir of excludeDirs) {
-      if (isDirectoryOrSubdirectory(excludeDir, p)) {
-        continue toNextProject
-      }
-    }
-
-    const projectDir = path.dirname(p)
-
-    /*
-      The project's directory might have been deleted as a result of a previous
-      cleanup step
-    */
-    if (!(await fsExists(projectDir))) {
-      continue
-    }
-
-    try {
-      results.push(
-        await executor(
-          "sh",
-          ["-c", "git add . && git reset --hard && git clean -xdf"],
-          { options: { cwd: projectDir } },
-        ),
-      )
-    } catch (error) {
-      results.push(error instanceof Error ? error : new Error(String(error)))
-    }
-  }
-
-  return results
-}
-
-const isDirectoryOrSubdirectory = (parent: string, child: string) => {
-  if (arePathsEqual(parent, child)) {
-    return true
-  }
-
-  const relativePath = path.relative(parent, child)
-  if (
-    relativePath &&
-    !relativePath.startsWith("..") &&
-    !path.isAbsolute(relativePath)
-  ) {
-    return true
-  }
-
-  return false
-}
-
-const arePathsEqual = (a: string, b: string) => {
-  return a === b || normalizePath(a) === normalizePath(b)
-}
-
 let lastIncrementalId = 0
 export const getNextUniqueIncrementalId = () => {
   const nextIncrementalId = ++lastIncrementalId
@@ -233,24 +108,6 @@ export const getNextUniqueIncrementalId = () => {
     "getNextUniqueIncrementalId overflowed the next ID",
   )
   return nextIncrementalId
-}
-
-const normalizePath = (v: string) => {
-  for (const [pattern, replacement] of [
-    [/\\/g, "/"],
-    [/(\w):/, "/$1"],
-    [/(\w+)\/\.\.\/?/g, ""],
-    [/^\.\//, ""],
-    [/\/\.\//, "/"],
-    [/\/\.$/, ""],
-    [/\/$/, ""],
-  ] as const) {
-    while (pattern.test(v)) {
-      v = v.replace(pattern, replacement)
-    }
-  }
-
-  return v
 }
 
 export const redact = (str: string, items: string[], replacement: string) => {
